Add Startup Version Check and Service Naming
