+++ bootstrap/peer.go
@@ -94,7 +94,7 @@ func New(log *zap.Logger, full *identity.FullIdentity, db DB, config Config, ver
 	var err error
 
 	{
-		peer.Version = version.NewService(&config.Version, &versionInfo)
+		peer.Version = version.NewService(&config.Version, &versionInfo, "Bootstrap")
 	}
 
 	{ // setup listener and server
