+++ internal/version/service.go
@@ -7,17 +7,14 @@ import (
 	"context"
 	"encoding/json"
 	"net/http"
+	"reflect"
 	"sync"
 	"time"
 
 	"github.com/zeebo/errs"
 	"go.uber.org/zap"
-	"google.golang.org/grpc"
 
 	"storj.io/storj/internal/sync2"
-	"storj.io/storj/pkg/identity"
-	"storj.io/storj/pkg/pb"
-	"storj.io/storj/pkg/transport"
 )
 
 // Config contains the necessary Information to check the Software Version
@@ -29,8 +26,9 @@ type Config struct {
 
 // Service contains the information and variables to ensure the Software is up to date
 type Service struct {
-	config *Config
-	info   *Info
+	config  *Config
+	info    *Info
+	service string
 
 	Loop *sync2.Cycle
 
@@ -39,71 +37,21 @@ type Service struct {
 	allowed bool
 }
 
-type VersionedClient struct {
-	transport transport.Client
-	version   *Service
-}
-
-const (
-	ErrOldVersion = "Outdated Software Version, please update!"
-)
-
 // NewService creates a Version Check Client with default configuration
-func NewService(config *Config, info *Info) (client *Service) {
+func NewService(config *Config, info *Info, service string) (client *Service) {
 	return &Service{
 		config:  config,
 		info:    info,
+		service: service,
 		Loop:    sync2.NewCycle(config.CheckInterval),
 		checked: make(chan struct{}, 0),
 		allowed: false,
 	}
 }
 
-// NewVersionedClient returns a transport client which ensures, that the software is up to date
-func NewVersionedClient(transport transport.Client, service *Service) *VersionedClient {
-	return &VersionedClient{
-		transport: transport,
-		version:   service,
-	}
-}
-
-// DialNode returns a grpc connection with tls to a node.
-//
-// Use this method for communicating with nodes as it is more secure than
-// DialAddress. The connection will be established successfully only if the
-// target node has the private key for the requested node ID.
-func (client *VersionedClient) DialNode(ctx context.Context, node *pb.Node, opts ...grpc.DialOption) (*grpc.ClientConn, error) {
-	if !client.version.IsUpToDate() {
-		return nil, errs.New(ErrOldVersion)
-	}
-	return client.transport.DialNode(ctx, node, opts...)
-}
-
-// DialAddress returns a grpc connection with tls to an IP address.
-//
-// Do not use this method unless having a good reason. In most cases DialNode
-// should be used for communicating with nodes as it is more secure than
-// DialAddress.
-func (client *VersionedClient) DialAddress(ctx context.Context, address string, opts ...grpc.DialOption) (*grpc.ClientConn, error) {
-	if !client.version.IsUpToDate() {
-		return nil, errs.New(ErrOldVersion)
-	}
-	return client.transport.DialAddress(ctx, address, opts...)
-}
-
-// Identity is a getter for the transport's identity
-func (client *VersionedClient) Identity() *identity.FullIdentity {
-	return client.transport.Identity()
-}
-
-// WithObservers returns a new transport including the listed observers.
-func (client *VersionedClient) WithObservers(obs ...transport.Observer) transport.Client {
-	return &VersionedClient{client.transport.WithObservers(obs...), client.version}
-}
-
 // Run logs the current version information
 func (srv *Service) Run(ctx context.Context) error {
-	firstCheck := false
+	firstCheck := true
 	return srv.Loop.Run(ctx, func(ctx context.Context) error {
 		var err error
 		allowed, err := srv.checkVersion(ctx)
@@ -115,6 +63,9 @@ func (srv *Service) Run(ctx context.Context) error {
 		if firstCheck {
 			close(srv.checked)
 			firstCheck = false
+			if !allowed {
+				zap.S().Fatal(ErrOldVersion)
+			}
 		}
 
 		if err != nil {
@@ -145,8 +96,7 @@ func (srv *Service) checkVersion(ctx context.Context) (allowed bool, err error) 
 
 	zap.S().Debugf("allowed versions from Control Server: %v", accepted)
 
-	// ToDo: Fetch own Service Tag to compare correctly!
-	list := accepted.Storagenode
+	list := getFieldString(&accepted, srv.service)
 	if list == nil {
 		return true, errs.New("Empty List from Versioning Server")
 	}
@@ -200,4 +150,14 @@ func (srv *Service) DebugHandler(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		zap.S().Errorf("error writing data to client %v", err)
 	}
+}
+
+func getFieldString(array *AllowedVersions, field string) []SemVer {
+	r := reflect.ValueOf(array)
+	f := reflect.Indirect(r).FieldByName(field).Interface()
+	result, ok := f.([]SemVer)
+	if ok {
+		return result
+	}
+	return nil
 }
